Make publish return true on success
The publish method will now return true when the provided namespace has
been resolved to an existing channel. This will make it easier to
determine if the call resulted in the desired result. Before this change
the method would return `null` if the namespace could not be resolved
and `undefined` if the message was passed along to the channel. By making
the positive branch return `true` there is less of a change to make a
false positive.

Just because the method returns `true` does not mean the message actually
was send out to any subscribers. This could be prevented by a predicate
on the subscribers.